Fix statuses not updating
Fixes https://github.com/singer-io/tap-braintree/issues/6

Issue #6 outlines the problem, summarized here.

Braintree transaction's API doesn't have a mechanism for pulling
transactions that were updated since a certain datetime. Transactions
API has the following status enum `authorized | submitted_for_settlement
| settling | settled`.

When using the current version of this tap, transactions are loaded a
single time on the timer interval (30 - 60min). By the time a
transaction is pulled from API it is often in the
`submitted_for_settlement` status. But due to lack of API support for
updated_at querying, there is no mechanism to update those rows, leading
to the transaction forever remaining as `submitted_for_settlment` in the
target.

After one or more hours pass, the API will register this transaction as
`settled`.

In order to easily have accurate data for status, we can continuously
fetch the trailing 3 days worth of data. This is the solution included
in this commit.

The approaches considered and otherwise rejected:
- Keep track of updated_at and created_at values in the state.json.
  - I observed that updated_at and created_at are identical in the DB
  row output and thus invalid for this use-case
- Use a smaller trailing window of time
  - This could be valid but without confirmation about timeframe in
  Braintree docs, it makes sense to keep a larger value

Benefits to this approach:
- Having status that reflects the data in Braintree

Drawbacks:
- Wastful with queries and rows sent through StitchData
  - This could be resolved by storing transaction_id, status in the
  state.json and only updating the row if it differs.
  - This adds much more code to manage, write and invalidate state.json
  info.
  - I don't see a strong need yet and would rather deal with a bit of
  wasteful usage.

So my proposal is to merge this PR and have an incremental improvement
from where this tap currently sits :).